Created index page and recipe page/folder
The index page contains the heading Odin Recipes and a link to the
macaroni recipe page, which remains blank

Explain the problem that this commit is solving. Focus on why you
are making this change as opposed to how (the code explains that).
Are there side effects or other unintuitive consequences of this
change? Here's the place to explain them.