Created index and recipe pages/folders
The index page includes a heading that says Odin Recipes with recipes
underneath in subheadings. When writing these pages, I sought a way
in VS Code to implement HTML boilerplate with a shortcut. VS Code
includes a shortcut that you can type (HTML:5). I was curious about the
meta tag with the attribute string "X-UA-Compatible" and learned that it
is for compatibility with IE browsers before version 11.