Updated style attributes to simplify code
Originally, the html tag held the attribute to style text as white, and
the body tag held the attribute to style background as grey. The style
attribute can hold two values if you put a semicolon between them.
The body style attribute was added to the html tag for simplification.